works with new api

+++ src/what/bookmarks/BookmarksActivity.java
@@ -53,7 +53,7 @@ public class BookmarksActivity extends MyActivity implements OnClickListener {
 	private void openBookmark(int i) {
 		Bundle b = new Bundle();
 		intent = new Intent(BookmarksActivity.this, what.torrents.torrents.TorrentTabActivity.class);
-		b.putInt("torrentGroupId", bookmarks.getResponse().getBookmarksList().get(i).getId());
+		b.putInt("torrentGroupId", bookmarks.getResponse().getBookmarksList().get(i).getId().intValue());
 		intent.putExtras(b);
 		startActivityForResult(intent, 0);
 	}
